stops at user agent sniffing
