+++ client/libs/startup.js
@@ -1,4 +1,4 @@
-import { User } from './cookies.js'
+import { User } from './incoming.js'
 
 Meteor.startup(() => {
   // code to run on server at startup
